refactor(CalculatorController): Controller 형태로 변경
Controller 역할에 집중된 비즈니스를 옮겨 리팩토링했습니다.
