+++ src/main/java/calculator/Application.java
@@ -1,10 +1,10 @@
 package calculator;
 
-import calculator.controller.CalculatorProcess;
+import calculator.controller.CalculatorController;
 
 public class Application {
     public static void main(String[] args) {
-        CalculatorProcess calculatorController = new CalculatorProcess();
+        CalculatorController calculatorController = new CalculatorController();
         calculatorController.run();
     }
 }
